Disable input during predictive back animation
