+++ extensions-compose/src/commonMain/kotlin/com/arkivanov/decompose/extensions/compose/stack/animation/AbstractStackAnimation.kt
@@ -8,9 +8,9 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.input.pointer.pointerInput
 import com.arkivanov.decompose.Child
 import com.arkivanov.decompose.ExperimentalDecomposeApi
+import com.arkivanov.decompose.extensions.compose.utils.InputConsumingOverlay
 import com.arkivanov.decompose.router.stack.ChildStack
 
 @OptIn(ExperimentalDecomposeApi::class)
@@ -56,23 +56,9 @@ internal abstract class AbstractStackAnimation<C : Any, T : Any>(
             // A workaround until https://issuetracker.google.com/issues/214231672.
             // Normally only the exiting child should be disabled.
             if (disableInputDuringAnimation && (items.size > 1)) {
-                Overlay(modifier = Modifier.matchParentSize())
+                InputConsumingOverlay(modifier = Modifier.matchParentSize())
             }
         }
-    }
-
-    @Composable
-    private fun Overlay(modifier: Modifier) {
-        Box(
-            modifier = modifier.pointerInput(Unit) {
-                awaitPointerEventScope {
-                    while (true) {
-                        val event = awaitPointerEvent()
-                        event.changes.forEach { it.consume() }
-                    }
-                }
-            }
-        )
     }
 
     private fun getAnimationItems(newStack: ChildStack<C, T>, oldStack: ChildStack<C, T>?): Map<Any, AnimationItem<C, T>> =
